Move HTTP header field name handling into separate unit, and extend BasicHttpHeaders
